+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 ### Changed
 
 - The `urlwatch` script (Git only) now works when run from different paths
+- Chunking of strings (e.g. for Slack and Telegram) now adds numbering (e.g.
+  ` (1/2)`) to the messages (only if a message is split into multiple parts)
 
 ## [2.19] -- 2020-07-17